@@ -6,15 +6,13 @@
     "authors": [
         {
             "name": "Michael Bodnarchuk",
-            "email": "davert.php@mailican.com"
+            "email": "davert@codeception.com"
         }
     ],
     "require": {
-        "php": ">=5.4.0",
-        "myclabs/deep-copy": "~1.1"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "~4.0"
+        "php": ">=7.0.0",
+        "myclabs/deep-copy": "~1.1",
+        "phpunit/phpunit": "~6.0"
     },
     "autoload": {
         "psr-0": {
