@@ -1,55 +1,79 @@
 <?php
 namespace Codeception;
 
-use Codeception\Specify\Config;
-use Codeception\Specify\ConfigBuilder;
+use Codeception\Specify\SpecifyTest;
 use Codeception\Specify\ObjectProperty;
+use PHPUnit\Framework\AssertionFailedError;
+use PHPUnit\Framework\TestCase;
 
 trait Specify
 {
-
     private $beforeSpecify = array();
     private $afterSpecify = array();
-
-    /**
-     * @var Specify\Config
-     */
-    private $specifyConfig;
 
     /**
      * @var \DeepCopy\DeepCopy()
      */
     private $copier;
 
-    private function specifyInit()
+    /**
+     * @var SpecifyTest
+     */
+    private $currentSpecifyTest;
+
+    private $specifyName = '';
+
+    /**
+     * @return SpecifyTest
+     */
+    public function getCurrentSpecifyTest()
     {
-        if ($this->copier) return;
-        $this->copier = new \DeepCopy\DeepCopy();
-        $this->copier->skipUncloneable();
-        if (!$this->specifyConfig) $this->specifyConfig = Config::create();
+        return $this->currentSpecifyTest;
     }
 
-    function specify($specification, \Closure $callable = null, $params = [])
+    public function should($specification, \Closure $callable = null, $params = [])
     {
-        if (!$callable) return;
-        $this->specifyInit();
+        $this->specify("should " . $specification, $callable, $params);
+    }
 
-        $test = $callable->bindTo($this);
-        $oldName = $this->getName();
-        $newName = $oldName . ' | ' . $specification;
+    public function it($specification, \Closure $callable = null, $params = [])
+    {
+        $this->specify("should " . $specification, $callable, $params);
+    }
 
-        $this->setName($newName);
+    public function describe($specification, \Closure $callable = null)
+    {
+        $this->specify($specification, $callable);
+    }
+
+    public function specify($specification, \Closure $callable = null, $params = [])
+    {
+        if (!$callable) {
+            return;
+        }
+
+        /** @var $this TestCase  **/
+        if (!$this->copier) {
+            $this->copier = new \DeepCopy\DeepCopy();
+            $this->copier->skipUncloneable();
+        }
 
         $properties = $this->getSpecifyObjectProperties();
 
         // prepare for execution
-        $throws = $this->getSpecifyExpectedException($params);
         $examples = $this->getSpecifyExamples($params);
         $showExamplesIndex = $examples !== [[]];
 
+        $specifyName = $this->specifyName;
+        $this->specifyName .= ' ' . $specification;
+
         foreach ($examples as $idx => $example) {
+            $test = new SpecifyTest($callable->bindTo($this));
+            $this->currentSpecifyTest = $test;
+            $test->setName($this->getName() . ' |' . $this->specifyName);
+            $test->setExample($example);
             if ($showExamplesIndex) {
-                $this->setName($newName . ' | examples index ' . $idx);
+                $test->setName($this->getName() . ' |' . $this->specifyName . ' # example ' . $idx);
             }
 
             // copy current object properties
@@ -61,7 +85,8 @@ trait Specify
                 }
             }
 
-            $this->specifyExecute($test, $throws, $example);
+            $test->run($this->getTestResultObject());
+            $this->specifyCheckMockObjects();
 
             // restore object properties
             $this->specifyRestoreProperties($properties);
@@ -73,8 +98,8 @@ trait Specify
             }
         }
 
-        // restore test name
-        $this->setName($oldName);
+        // revert specify name
+        $this->specifyName = $specifyName;
     }
 
     /**
@@ -91,79 +116,23 @@ trait Specify
         return [[]];
     }
 
-    private function getSpecifyExpectedException($params)
+    /**
+     * @return \ReflectionClass|null
+     */
+    private function specifyGetPhpUnitReflection()
     {
-        if (isset($params['throws'])) {
-            $throws = (is_array($params['throws'])) ? $params['throws'][0] : $params['throws'];
-
-            if (is_object($throws)) {
-                $throws = get_class($throws);
-            }
-            if ($throws === 'fail') {
-                $throws = 'PHPUnit_Framework_AssertionFailedError';
-            }
-
-            $message = (is_array($params['throws']) && isset($params['throws'][1])) ? $params['throws'][1] : false;
-
-            return [$throws, $message];
-        }
-
-        return false;
-    }
-
-    private function specifyExecute($test, $throws = false, $examples = array())
-    {
-        $message = false;
-
-        if (is_array($throws)) {
-            $message = ($throws[1]) ? strtolower($throws[1]) : false;
-            $throws = $throws[0];
-        }
-
-        $result = $this->getTestResultObject();
-
-        try {
-            call_user_func_array($test, $examples);
-            $this->specifyCheckMockObjects();
-        } catch (\PHPUnit_Framework_AssertionFailedError $e) {
-            if ($throws !== get_class($e)){
-                $result->addFailure(clone($this), $e, $result->time());
-            }
-
-            if ($message !==false && $message !== strtolower($e->getMessage())) {
-                $f = new \PHPUnit_Framework_AssertionFailedError("exception message '$message' was expected, but '" . $e->getMessage() . "' was received");
-                $result->addFailure(clone($this), $f, $result->time());
-            }
-        } catch (\Exception $e) {
-            if ($throws) {
-                if ($throws !== get_class($e)) {
-                    $f = new \PHPUnit_Framework_AssertionFailedError("exception '$throws' was expected, but " . get_class($e) . ' was thrown');
-                    $result->addFailure(clone($this), $f, $result->time());
-                }
-
-                if ($message !==false && $message !== strtolower($e->getMessage())) {
-                    $f = new \PHPUnit_Framework_AssertionFailedError("exception message '$message' was expected, but '" . $e->getMessage() . "' was received");
-                    $result->addFailure(clone($this), $f, $result->time());
-                }
-            } else {
-                throw $e;
-            }
-        }
-
-        if ($throws) {
-            if (isset($e)) {
-                $this->assertTrue(true, 'exception handled');
-            } else {
-                $f = new \PHPUnit_Framework_AssertionFailedError("exception '$throws' was not thrown as expected");
-                $result->addFailure(clone($this), $f, $result->time());
-            }
+        if ($this instanceof \PHPUnit\Framework\TestCase) {
+            return new \ReflectionClass(\PHPUnit\Framework\TestCase::class);
         }
     }
 
-    public function specifyConfig()
+    private function specifyCheckMockObjects()
     {
-        if (!$this->specifyConfig) $this->specifyConfig = Config::create();
-        return new ConfigBuilder($this->specifyConfig);
+        if (($phpUnitReflection = $this->specifyGetPhpUnitReflection()) !== null) {
+            $verifyMockObjects = $phpUnitReflection->getMethod('verifyMockObjects');
+            $verifyMockObjects->setAccessible(true);
+            $verifyMockObjects->invoke($this);
+        }
     }
 
     function beforeSpecify(\Closure $callable = null)
@@ -184,33 +153,6 @@ trait Specify
     /**
      * @param ObjectProperty[] $properties
      */
-    private function specifyCloneProperties($properties)
-    {
-        foreach ($properties as $property) {
-            $propertyName = $property->getName();
-            $propertyValue = $property->getValue();
-
-            if ($this->specifyConfig->classIgnored($propertyValue)) {
-                continue;
-            }
-
-            if ($this->specifyConfig->propertyIsShallowCloned($propertyName)) {
-                if (is_object($propertyValue)) {
-                    $property->setValue(clone $propertyValue);
-                } else {
-                    $property->setValue($propertyValue);
-                }
-            }
-
-            if ($this->specifyConfig->propertyIsDeeplyCloned($propertyName)) {
-                $property->setValue($this->copier->copy($propertyValue));
-            }
-        }
-    }
-
-    /**
-     * @param ObjectProperty[] $properties
-     */
     private function specifyRestoreProperties($properties)
     {
         foreach ($properties as $property) {
@@ -224,42 +166,35 @@ trait Specify
     private function getSpecifyObjectProperties()
     {
         $objectReflection = new \ReflectionObject($this);
-        $propertiesToClone = $objectReflection->getProperties();
+        $properties = $objectReflection->getProperties();
 
         if (($classProperties = $this->specifyGetClassPrivateProperties()) !== []) {
-            $propertiesToClone = array_merge($propertiesToClone, $classProperties);
+            $properties = array_merge($properties, $classProperties);
         }
 
-        $properties = [];
+        $clonedProperties = [];
 
-        foreach ($propertiesToClone as $property) {
-            if ($this->specifyConfig->propertyIgnored($property->getName())) {
+        foreach ($properties as $property) {
+            /** @var $property \ReflectionProperty  **/
+            $docBlock = $property->getDocComment();
+            if (!$docBlock) {
                 continue;
             }
-
-            $properties[] = new ObjectProperty($this, $property);
+            if (preg_match('~\*(\s+)?@specify\s?~', $docBlock)) {
+                $property->setAccessible(true);
+                $clonedProperties[] = new ObjectProperty($this, $property);
+            }
         }
 
-        // isolate mockObjects property from PHPUnit_Framework_TestCase
-        if (($phpUnitReflection = $this->specifyGetPhpUnitReflection()) !== null) {
-            $properties[] = $mockObjects = new ObjectProperty(
-                $this, $phpUnitReflection->getProperty('mockObjects')
-            );
-
+        // isolate mockObjects property from PHPUnit\Framework\TestCase
+        if ($classReflection = $this->specifyGetPhpUnitReflection()) {
+            $property = $classReflection->getProperty('mockObjects');
             // remove all mock objects inherited from parent scope(s)
-            $mockObjects->setValue([]);
+            $clonedProperties[] = new ObjectProperty($this, $property);
+            $property->setValue($this, []);
         }
 
-        return $properties;
-    }
-
-    private function specifyCheckMockObjects()
-    {
-        if (($phpUnitReflection = $this->specifyGetPhpUnitReflection()) !== null) {
-            $verifyMockObjects = $phpUnitReflection->getMethod('verifyMockObjects');
-            $verifyMockObjects->setAccessible(true);
-            $verifyMockObjects->invoke($this);
-        }
+        return $clonedProperties;
     }
 
     private function specifyGetClassPrivateProperties()
@@ -277,12 +212,13 @@ trait Specify
     }
 
     /**
-     * @return \ReflectionClass|null
+     * @param ObjectProperty[] $properties
      */
-    private function specifyGetPhpUnitReflection()
+    private function specifyCloneProperties($properties)
     {
-        if ($this instanceof \PHPUnit_Framework_TestCase) {
-            return new \ReflectionClass('\PHPUnit_Framework_TestCase');
+        foreach ($properties as $property) {
+            $propertyValue = $property->getValue();
+            $property->setValue($this->copier->copy($propertyValue));
         }
     }
 }
